Converted to float16

+++ src/env_ca.py
@@ -72,7 +72,7 @@ class Env_CA(nn.Module):
         kernel = kernel_stack.unsqueeze(1)
         
         # Repeat kernels to form num_channels x 1 x 3 x 3 filter
-        kernel = kernel.repeat(self.num_channels, 1, 1, 1).float().to(self.device)
+        kernel = kernel.repeat(self.num_channels, 1, 1, 1).half().to(self.device)
 
         state_repeated = state_grid.repeat_interleave(kernel_stack.shape[0],dim = 1)
